nts: Revert to int64_t [skip ci]
Some floats are wider than 32 int32_t

+++ src/lfortran/semantics/ast_body_visitor.cpp
@@ -1475,7 +1475,7 @@ public:
                                 if (LFortran::ASR::is_a<LFortran::ASR::Real_t>(*func_type)) {
                                     if (func_kind == 4){
                                         float rr = ASR::down_cast<ASR::ConstantReal_t>(LFortran::ASRUtils::expr_value(func_expr))->m_r;
-                                        int32_t ival = static_cast<int32_t>(rr);
+                                        int64_t ival = static_cast<int64_t>(rr);
                                         value = ASR::down_cast<ASR::expr_t>(ASR::make_ConstantInteger_t(al, x.base.base.loc, ival, func_type));
                                     } else {
                                         double rr = ASR::down_cast<ASR::ConstantReal_t>(LFortran::ASRUtils::expr_value(func_expr))->m_r;
@@ -1485,7 +1485,7 @@ public:
                                 }
                                 else if (LFortran::ASR::is_a<LFortran::ASR::Integer_t>(*func_type)) {
                                     if (func_kind == 4){
-                                        int32_t ival = ASR::down_cast<ASR::ConstantInteger_t>(
+                                        int64_t ival = ASR::down_cast<ASR::ConstantInteger_t>(
                                             LFortran::ASRUtils::expr_value(func_expr))->m_n;
                                         value = ASR::down_cast<ASR::expr_t>(ASR::make_ConstantInteger_t(al, x.base.base.loc, ival, func_type));
                                     } else {
@@ -1517,7 +1517,7 @@ public:
                                 }
                                 else if (LFortran::ASR::is_a<LFortran::ASR::Integer_t>(*func_type)) {
                                     if (real_kind == 4){
-                                        int32_t rv = ASR::down_cast<ASR::ConstantInteger_t>(
+                                        int64_t rv = ASR::down_cast<ASR::ConstantInteger_t>(
                                             LFortran::ASRUtils::expr_value(real_expr))->m_n;
                                         float rr = static_cast<float>(rv);
                                         value = ASR::down_cast<ASR::expr_t>(ASR::make_ConstantReal_t(al, x.base.base.loc, rr, func_type));
